move request handlers for columns to separate file
