+++ server/index.js
@@ -4,6 +4,8 @@ const express = require('express');
 
 const errorMiddleware = require('./error-middleware');
 const staticMiddleware = require('./static-middleware');
+const columns = require('./controllers/columns');
+
 const db = new pg.Pool({
   connectionString: process.env.DATABASE_URL,
   ssl: {
@@ -17,6 +19,7 @@ app.use(staticMiddleware);
 
 app.use(errorMiddleware);
 // boards
+
 app.get('/api/users/:id/boards', (req, res) => {
   const { id } = req.params;
   const sql = `
@@ -87,58 +90,20 @@ app.put('/api/users/:id/boards/:boardId', async (req, res) => {
 
 // columns
 
-app.get('/api/users/:id/boards/:boardId/col', async (req, res) => {
-  const { boardId } = req.params;
-  const sql = `
-    SELECT *
-      FROM "columns"
-    WHERE "boardId" = $1
-  `;
-  const values = [boardId];
-  const result = await db.query(sql, values);
-  const data = result.rows;
-  res.json(data);
+app.get('/api/users/:id/boards/:boardId/col', (req, res) => {
+  columns.get(req, res, db);
 });
 
-app.post('/api/users/:id/boards/:boardId/col', async (req, res) => {
-  const sql = `
-  INSERT INTO "columns" ("boardId", name)
-    VALUES (12, 'New Column')
-    RETURNING *
-`;
-  const results = await db.query(sql);
-  const [data] = results.rows;
-  res.status(201).json(data);
+app.post('/api/users/:id/boards/:boardId/col', (req, res) => {
+  columns.create(req, res, db);
 });
 
-app.delete('/api/users/:id/boards/:boardId/col/:colId', async (req, res) => {
-  const { colId } = req.params;
-  const sql = `
-    DELETE from "columns"
-      WHERE "columnId" = $1
-  `;
-  const val = [colId];
-  await db.query(sql, val);
-  res.sendStatus(204);
+app.delete('/api/users/:id/boards/:boardId/col/:colId', (req, res) => {
+  columns.delete(req, res, db);
 });
 
 app.put('/api/users/:id/boards/:boardId/col/:colId', async (req, res) => {
-  const { colId } = req.params;
-  const sql = `
-    UPDATE "columns"
-        SET "name" = $1
-      WHERE "columnId" = $2
-    RETURNING *
-  `;
-  const values = [req.body.name, colId];
-  try {
-    const result = await db.query(sql, values);
-    const [record] = result.rows;
-    res.json(record);
-  } catch (err) {
-    console.error('error:', err.message);
-    res.status(500).send({ error: 'something went wrong' });
-  }
+  columns.edit(req, res, db);
 });
 
 app.listen(process.env.PORT, () => {
